pyproject.toml
[tool.poetry]
name = "gh-pr-release"
version = "0.1.0"
description = ""
authors = ["aita <792803+aita@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.7"
gitpython = "^2.1"
pygithub = "^1.43"
click = "^7.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
flake8 = "^3.7"
black = "==19.3b0"
mypy = "^0.720.0"

[tool.poetry.scripts]
gh-pr-release = 'gh_pr_release.cli:main'
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


